BIG-16046 Support non-remote endpoints for request
