+++ src/api/countries.js
@@ -1,7 +1,7 @@
-import Base from './base';
+import Remote from './remote';
 import Hooks from '../hooks';
 
-export default class extends Base
+export default class extends Remote
 {
     /**
      * @Constructor
