+++ src/api/search.js
@@ -1,8 +1,8 @@
 import _ from 'lodash';
-import Base from './base';
+import Remote from './remote';
 import Hooks from '../hooks';
 
-export default class extends Base
+export default class extends Remote
 {
     /**
      * @Constructor
